requirements.txt
apache-kafka==2.13.0
pyspark==3.3.1
opencv-python==4.7.0.72
numpy==1.23.5
tensorflow==2.12.0
yolo==0.1.2
psycopg2==2.9.6
schedule==1.1.0
